feat: fix #69 add agentVersion on host and backup
This commit adds the `agentVersion` field to the `AuthenticateReply` message in the `woodstock.proto` file and the `Backup` DTO in the `backup.dto.ts` file. The `agentVersion` field is used to store the version of the agent. This change allows for better tracking and identification of the agent version in the system.